+++ packages/cli/src/commands/add/__tests__/add.ts
@@ -4,7 +4,11 @@ import * as git from "@changesets/git";
 import { defaultConfig } from "@changesets/config";
 import { temporarilySilenceLogs } from "@changesets/test-utils";
 
-import { askCheckboxPlus, askConfirm, askQuestion } from "../../../utils/cli";
+import {
+  askCheckboxPlus,
+  askConfirm,
+  askQuestion
+} from "../../../utils/cli-utilities";
 import addChangeset from "..";
 import writeChangeset from "../writeChangeset";
